142일 25-01-22 JlModel.php update,delete primary이란 값이 primary 컬럼명에 대조해서 들어가게 로직추가
JlService.php
모델에 로직 추가됨으로 모델함수 실행시키기전 primary값 변환시키던 로직 삭제
delete() getData 조건 setFilter로 변경

JlVue.js
postData() 전역으로 사용할수있는 insert,update 템플릿 통신함수 추가
getData() 전역으로 사용할수있는 단일 데이터가져오는 템플릿 통신함수 추가
getsData() 전역으로 사용할수있는 복수형 데이터가져오는 템플릿 통신함수 추가
deleteData() 전역으로 사용할수있는 delete 템플릿 통신함수 추가

detail.php
jl 전역 통신함수 상세 사용법 추가

simple.php
템플릿 통신함수 삭제 및 기본 데이터셋 변경

+++ 장동혁템플릿/vue3/detail.php
@@ -52,6 +52,14 @@
                     page : 1,
                     limit : 1,
                     count : 0,
+                },
+
+                required : [
+                    {name : "",message : ``},
+                ],
+
+                data : {
+                    table : "",
                 }
             };
         },
@@ -182,6 +190,7 @@
                 }
             },
             async deleteData() {
+                if(! await this.jl.confirm("정말 삭제하시겠습니까?")) return false;
                 let method = "delete";
                 //let method = "where_delete";
 
@@ -219,6 +228,46 @@
                     alert(e.message)
                 }
             },
+
+            async jlPostData() {
+                let data = {
+                    table : "",
+                    name : "",
+                }
+
+                let required = [
+                    {name : "",message : ``},
+                ],
+                this.jl.postData(data,required,async (res) => {
+                    // 세번째 매개변수인 콜백함수가 빈값일 경우 밑에 코드가 자동 실행됌
+                    await this.jl.alert("완료되었습니다.");
+                    window.location.reload();
+                });
+            }
+
+            async jlGetData() {
+                let filter = {
+                    table : "",
+                    primary : "",
+                }
+                let data = this.jlGetData(filter,async (res) => {
+                    //함수내용
+                    // 두번쨰 매개변수인 콜백함수가 빈값일경우 res.data[0]이 그냥 return 됌
+                });
+            }
+
+            async jlGetsData() {
+                let filter = {
+                    table : "",
+                    primary : "",
+                }
+
+                let datas = this.jlGetsData(filter,async (res) => {
+                    //두번쨰 매개변수인 콜백함수가 빈값일경우 밑에 함수가 자동으로 적용됌
+                    this.data_array = res.data
+                    this.paging.count = res.count;
+                });
+            }
         },
         computed: {
